Fixed bugs in hard decoder

+++ hw4/bit_to_symbol_mapper.py
@@ -12,6 +12,6 @@ def bit_to_symbol_mapper(bits: np.ndarray, d: float = 1) -> np.ndarray:
 
 def symbol_to_bit_mapper(symbols: np.ndarray, d: float = 1) -> np.ndarray:
     assert symbols.dtype == np.complex128, "symbols must be complex"
-    i_bits = symbols.real + (d/2)
-    q_bits = symbols.imag + (d/2)
+    i_bits = (symbols.real + (d/2)).astype(int)
+    q_bits = (symbols.imag + (d/2)).astype(int)
     return np.ravel(np.column_stack((i_bits, q_bits)))
